+++ src/subsearch/scraper/opensubtitles.py
@@ -1,5 +1,5 @@
-from src.utilities import local_paths, log
-from src.scraper import opensubtitles_soup
+from util import local_paths, log
+from scraper import opensubtitles_soup
 
 
 # decides what to do with all the scrape data
